requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pymongo==4.12.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
mongoengine==0.27.0